split types into respective packages
